Make section titles lighter (and a bit bigger).
Since they don't contain real content, they don't need to stand out
that much. Lightening does that, and making the font a bit bigger makes
it a bit easier to read when you do want to.

Also, darkened the ">" in the breadcrumbs to show that it's not part
of the link.