Add Crossdart links
Added them to the _source_code partial, so every time we show the source
code for something, it will have the Crossdart link.

Also, added --add-crossdart cli option.
Unfortunately, I couldn't find any global config entity in DartDoc, and
passing addCrossdart bool through all the code down to SourceCodeMixin
seems to be too dirty and awkward.

So I created the Config class, which for now contains only one field
(addCrossdart), and use that class in SourceCodeMixin to decide whether
I should add crossdart url or not.

The protocol and domain for crossdart link is '//', so it'd match the
currently used protocol.