Make sorting better by ordering Uint8List before Uint16List.
Current sorting by name is plain lexical ordering on (upper-cased) code units.
This change will order embedded integers lexicograpically - longer numbers
sort after shorter numbers (prefixed zeroes are counted in the length).
This ensures that "8" is before "16", but doesn't equate "002" and "2".
Prefixed numbers are likely to be rare in practice, and when they occur, it's
usually because they are grouped with other numbers of the same length, which
sorts fine either way.

BUG=
R=keertip@google.com

Review URL: https://codereview.chromium.org//1419553004.