Add ability to "Crossdartify" the source code
If there's a crossdart.json file in the input dir (which can be
generated by Crossdart), it will use that file to add links to
crossdart.info in the source code block.

Since now we should support links in the source code section, I removed
`<code>` wrapping tags from it, and also now I HTML escape the source
code, and use triple mustache curly braces (`{{{sourceCode}}}`) to
insert it to the page.

Hopefully, that will simplify developer's life, when they want to look
into the method's implementation.