Fix crossdart link of inherited method from other packages
When there're inherited methods in a class, and they are inherited from
a class in other package (or SDK), we incorrectly built the link.

This fixes it, and now I try to infer the right path from `uri` and
`path` properties of `node.element.source`. If `uri` is not `dart:` or
`package:`, I just won't add the crossdart link (for safety).

I removed a test, because couldn't figure out how to make `package:`
`uri` in the tests - it's always `file:///` :(