`pub run test` fails if `dart` is not in your path
These also might fail if the `dart` in your path was
of a different version than you expected.  This happens
with Flutter since flutter downloads its own copy of the
Dart SDK which may be differnt from the one installed
globally on your disk.

In my case, I just was using an explicit path to `pub`
for `pub run test` and was surprised to see test failures.

@keertip