Add more ways where to locate crossdart.json files
This change is mostly for Flutter, to add crossdart links to
docs.flutter.io.

The way Flutter generates docs - it consists of several packages, and
then there's one 'artificial' project which imports all of those
packages. We run dartdoc on that "project", this way generating docs for
all the flutter packages together, so it looks like one package.

To support that for CrossDart support in DartDoc, I need to make some
changes. Before, we just looked for 'crossdart.json' file in the project
root, and if it's there - then use it to add CrossDart information to
the generated docs.

This commit adds looking for 'crossdart.json' files in dependent
packages as well. If we analyze a file during docs generation, and it's
from a different package, we check for a crossdart.json file in the root
of that package, and if it's there - use it for augmenting docs with
crossdart stuff.

Testing: Unit tests pass, I also deployed this version of dartdoc to
www.dartdocs.org and ran it there for several days to check if there
will be any errors.