Fix bug when parameterized typedef props have wrong return type [#1263]
@HansMuller noticed, that for properties, that have parameterized
typedef (i.e. with generic parameter) as a return type - they don't show
that parameter in the docs in their return types.

It appears we don't show them, because in that case
`FunctionType#typeFormals` return an empty array. If we use
`FunctionType#typeArguments`, then everything works fine.

After further investigation I figured that if we run dartdoc on the
whole Flutter codebase, `FunctionType#typeFormals` always returns empty
array. So, not sure why it was introduced in the first place.

`git blame` has shown it was introduced here
https://github.com/dart-lang/dartdoc/commit/9d96d8dbe24a15af8109a512bb463f235cfbebf5
by @keertip, as part of upgrade to the latest analyzer (0.27.0).

Though after the changes in this PR all unit tests are still green, and
also I couldn't find anything broken in the Flutter docs after quick
glance, I'm still a bit nervous about this change, because:

* I don't quite understand what's the difference between `typeFormals`
  and `typeArguments`, and in what cases we need to use `typeFormals`.
* I don't understand why that `if` condition with `typeFormals` was
  initially introduced.

So, I'd like to get advice from @keertip in that case, I think.