Rank parent class method higher than its overrides in search suggestions
We do that by saving `overriddenDepth` to index.json (which is just
how many overridden elements in the inheritance chain that element has),
and then reducing weight of elements in the suggestions according to
`overriddenDepth`.

Also, fixed `grind update-test-package-docs`, it didn't know where to
find examples for `@examples` directives.

Fixes #896