Deemphasize and resort the inherited members [#641]
@sethladd asked to reorder the list of methods and properties in a way
that the inherited members go last, and also deemphasized in the output
(the names will be nonbold and italic).

This way it will be easier to grasp when skimming through the docs what
are the new methods/fields of a class, but you still can see all of them
and we don't introduce any new togglers/switchers, which would
complicate the UI/UX.

Testing: Generated Flutter docs, made sure the inherited members are
nonbold/italic and go last.

Fixes #641