Add --auto-include-dependencies [#1159] (#1327)
If `--auto-include-dependencies` flag is provided, dartdoc tries to
automatically add all the used libraries, even from other packages, to
the list of the documented libraries.

For that, it goes through each model element in a package, and finds
"used" elements for each element (e.g. from the parameters and from the
return type), and if any of these are from undocumented lib - it adds
that lib, and then rebuild `Package` with the new list of libs.

Testing: Flutter now generates docs for `vector_math_64` when
`--auto-include-dependencies` are provided. And the test package
adds library from `test_package_imported` because it uses it and has one
of the members of that lib in one of the Apple method's parameters.

Fixes #1159