Add annotations to features (#1332)
This change addresses #1265 with a rearrangement of how the "features" section of italicized text is generated under each method or property.  Annotations, including custom annotations, are displayed, with the exception of "@override" and "@deprecated" as those are indicated by other means in dartdoc.  Final is now indicated instead of read-only for variables/properties declared final.