Search box shouldn't always use the first suggestion
When you select an item from the search box dropdown, we were processing
the event twice: once in typeahead:select and once on keydown. The
second handler was overwriting the window.location generated by the
first and always sending the user to the first suggestion even if they
had selected a different suggestion.

Fixes #1330