Add markdown renderers and renderer factory (#2152)
* Implement renderers for markdown

* Remove unnecessary strikethroughs in templates

* Add some tests

Only enough to mirror the few tests we have that exercise the existing
renderers. In the future, we should have a more comprehensive suite
that covers both html and md.

* dartfmt

* Call renderLinkedName directly from renderCategoryLabel

This avoids calling back to Category, which immediately calls the
renderer again anyway, except that during a test the second call went
to an html renderer because the packageGraph was holding a
HtmlRendererFactory.