Split apart ModelElement.from into a few pieces for cleaner code
ModelElement.from was huge; it handles all types of elements, including
property-inducing ones. It handled verification, and element caching.  This
changes splits it into 4 functions:

* ModelElement.fromPropertyInducingElement, a factory constructor.
* ModelElement.from, the sibling constructor for non-property-inducing
  elements.
* ModelElement._from, a helper which performs the actual constructions.
* ModelElement._cacheNewModelElement, a helper which caches the newly
  constructed ModelElement.

Callers frequently had conditional logic for property-inducing elements, so the
splitting of the constructor does not complicate callers, but allows them to be
explicit about which constructor they are calling.