Typedef implements ModelFunctionTyped.
The relationship is different from ModelFunctionTypedef which _extends_
ModelFunctionTyped because the typedef is the alias _which contains_ a
function-typed element. Nevertheless, Typedef was always implementing
(approximately) the interface defined by ModelFunctionTyped because both types
are sent to the _callable template.

Additionally one change is made to the ModelFunctionTyped interface such that
Typedef perfectly implements ModelFunctionTyped.

* **Breaking change**: Remove the `typeParameters` setter. The field is now final.