Introduce a Mustachio Builder.
This change is mostly concerned with gathering 'Renderer' annotations (which
isn't included yet; only a mock is included in specs), and walking the graph of
properties and types to gather all of the types which need renderers, for
resolving keys into property values.

There isn't much generated code in this change, but the base renderer is
included, which is referenced in generated renderer classes (stubs, for now).