Mustachio: Add property getters to Property, and end-to-end Renderer test
This includes some unorthodox tests: I've added a package in
`testing/mustachio`, complete with a pubspec.yaml and a build.yaml. These are
required in order to test the runtime behavior of generated code; in order to
make assertions on the generated code (not its String content, but its runtime
properties), it must be generated with public classes (the default is private).
In order to pass a configuration to the Builder, I need a build.yaml separate
from dartdoc's.

In order to import `testing/mustachio/lib/foo.renderers.dart`, and resolve that
library's reference to `package:testing_mustachio/foo.dart`, I need to add
`testing_mustachio` to `dev_dependencies`.

In order to reference the builders from the testing package, the builder code
needed to move from `tool/mustachio/` to `lib/src/mustachio/`.

The generated code now always references the library that generated it, on the
assumption that it needs types from that file. This is not necessarily correct,
but is a decent approximation until we start using code_builder.

RendererBase now needs to know about the renderer of the parent context type,
so it gets a new field.

Property and each renderer's propertyMap gets corrected code; the runtime tests
revealed some runtime type bugs.