Mustachio: Introduce Template class for pre-parsing template files. (#2484)
Mustachio: Introduce Template class for pre-parsing template files.

Second, change the signature of a partial resolver to return a Future<File>.
This allows for dartdoc's current partial parsing, which requires use of an
asynchronous Isolate API to get the path of a "package:" URI asset.

Before this change, Mustachio required re-parsing a template file for every
object which was to be rendered with that file. Horribly inefficient. This
change makes Mustachio much more similar to other Mustache parsers which allow
the user to handle a Template object which can be used to render multiple
context objects.

When parsing a Mustache template into a Template object, all partials
(recursive) will also be resolved, read, and parsed into Template objects. This
change has broad and beneficial consequences across the APIs which previously
passesd around Files and PartialResolvers:

* A RenderFunction now takes a Template instead of a File and a
  PartialResolver.
* RenderBase() now takes a Template instead of a File and a PartialResolver,
  and never concerns itself with partial resolution.
* RenderBase.partial now just retreives the pre-parsed partial instead of
  parsing inline.
* SimpleRenderer and renderSimple now take a Template instead of a File and a
  PartialResolver.

Also, the meta package is bumped in order to use `@internal`.