Mustachio: avoid calling Iterable getter multiple times (#2546)
Mustachio: avoid calling Iterable getter multiple times

Take this template, where `widgets` is an Iterable property:

Text {{#widgets}} {{name}} {{/widgets}}.

A mustache renderer should assume that the `widgets` getter may have side
effects, and it should only be called one time. Currently, Mustachio will call
this getter 2 or 3 times, as the behavior differs if the section is inverted,
and if the Iterable is empty.

This change is actually quite readable, and simplifies the code. The magic
sauce is in the lazy Iterable returned by `renderIterable`. If Mustachio does
not wish to render the elements, then the lazy map will not be evaluated.