Simplify logic in features rendering.
The default template calls `featuresAsString` twice, once to see if it is empty, then again to render the poor String. `featuresAsString` involves:

1. Iterating over all annotations,
2. finding canonical elements,
3. HTML-escaping the source of each one,
4. sorting the results into a list, by a custom sort.

This change introduces a new `hasFeatures` getter. Using this in templates:

1. avoids HTML-escaping,
2. avoids sorting,
3. short-circuits annotations if the element is `final` or `late`,
4. returns `true` when the first displayable element is found, sometimes avoiding iterating.

Additionally, the HtmlEscape is extracted into a static const.

A `realElement` getter is introduced on ElementAnnotation, which is used in a number of places.

Annotations are not HTML-escaped unless we definitely want to display them.