Modified arrow for fields to have different meaning when private.
The arrow was either → for fields with public getters, ← for fields with
public setters, and ↔ for fields with both.

With the inclusion of private fields, we need to account for a complete
lack of public getters and setters. When a field is private, we think of
mutability in terms of the field's library, where it is accessible.

In this case, → means only a private getter, ← means only a private setter,
and ↔ means both.