Replaced all instances of `Container.publicX` with `Container.documentedX`.
With the new ability to document private elements, being public is no longer the
deciding factor in whether to document something. This commit changes the
abstract getters in `Container`, and the following commits will change the
concrete implementations to use `Documentable.isDocumented` instead of
`Privacy.isDocumented`. Presumably, this will also get rid of all references to
`model_utils.filterNonPublic` in favor of `model_utils.filterNonDocumented`.

There are a lot of errors here until all the subclasses are changed as well.