Invert a type test on a type variable
This type test was introduced during the null safety migration and I do
not see discussion about the intention.
In practice this is `assert(true)` so it would be safest to remove it
entirely, but we can also take a guess at the intention.

The full assert when it was introduced was
`asset(T is! Iterable || splitCommas == true)`.

Later the second condition was removed, and the `splitCommas` field
overridden to always return `false`.

I think the intention was to express that options which support multiple
values _must_ split commas, and now that this class is only used without
splitting commas the intention is that the options must be scalar
values.

Invert the type test with a value known to satisfy a check for any
subtype of `Iterable`. Add a string explaining the assert.