Avoid Lists of MapEntries.
They are pretty hard to reason about, and also some "put all if absent" code
was employed to wrangle them. But it turns out that `a.putAllIfAbsent(b)` is
the same as `b.addAll(a)`. So this code can be simplified all around by just
using Maps as they are meant to be used.