In PackageBuilder._discoverLibraries, initialize newFiles outside loop. (#3695)
I noticed with print debugging that `addFilesReferencedBy` was being called way
too many times for a given library. It turns out this is because it was being
called on different Sets, when really it should mostly be operating on one Set.
That one Set is `newFiles`. It was initialized in the do-while loop, but it can
be initialized before the do-while loop.

I checked the logic to make sure this change is a no-op.

I also timed documenting flutter. This change seems to reduce the
time-to-document by ~100 seconds, but that is only with one before and one
after, so I promise nothing.