Remove the TypeImplementing mixin; unnecessary (#3705)
This mixin on `InheritingContainer` was mixed into all 3 types that extended
`InheritingContainer`, so it sort of just carved out some functionality. But I
think we've agreed that the amount of carving out has made dartdoc harder to
read. This mixin provided 2 fields and 7 getters (2 of which are overrides),
which I have just merged into `InheritingContainer, which I have just merged
into `InheritingContainer`.

I actually made it a little simpler as well, removing `publicImplementers`,
which was unnecessary, and renamed `interfaces` to `directInterfaces`, and made
that `@visibleForTesting`.

Also while working here I saw:

* `_isInheritingFrom` only ever takes a non-nullable; simplify that.