Convert many fields on Container and TopLevelContainer to getters (#3710)
These were late, final, and public, my least favorite thing. The idea,
historically, behind late final fields in dartdoc is caching. If a value is
expensive to calculate, like a big list of things, and we have to access that
value a few times, best to calculate it once.

But some late final fields are only accessed while rendering HTML, accessed
from the rendered templates. And typically only accessed once. The '*Sorted'
fields definitely fall into this bucket. So I made them all getters, and then
benchmarked with the googleapis package, and found no significant change in
time-to-document, or max RSS.