Rewrite the 'exported extension method' tests; add basic extension tests. (#3738)
I found this file and it was very hard for me to understand. It contained three
tests for extensions being exported from private to public API. I've preserved
the tests, but in an entirely new format. It is more standard, I think, and
more readable.

Since these were the only tests in a file called 'extension_methods_test', I
decided that I should add a few basic extension tests (and move the file to
extension_tests) to differentiate, a bit, what the file should contain.