Simplify Library.allModelElements
This was backed by another field called `modelElementsMap`, but that is now an
unnecessary complexity; so combine the field and getter into just a field.

Additionally, `PackageGraph.allLocalModelElements` is no longer needed. It was
used in tests, but those tests can calculate the list of all model elements for
themselves.

Also simplify `buildDocumentationAddition`, and improve doc comments.