Avoid Lists of MapEntries. (#3807)
In the existing code, a few APIs like
`CommentReferableEntryGenerators.generateEntries` and
`Container.extraReferenceChildren` used `List<MapEntry<...>>` which I've found
super confusing.

They are pretty hard to reason about, and also some "put all if absent" code
was employed to wrangle them. But it turns out that given Map `a` and Map `b`,
`a.putAllIfAbsent(b)` is the same as `b.addAll(a)`. So this code can be
simplified all around by just using Maps as they are meant to be used.

This code is a no-op; same results, just simpler.