Correct a regression for a package with a specified set of included libraries (#3825)
Correct a regression for a package with a specified set of included libraries

In 12d271a542c5e7ebb96763a66f560e2f4a99ca24 we change what libraries can count
as a 'canonical library', such that when a set of libraries is specified with
the 'include' option, any library not in the 'include' list is not canonical.
This breaks a few things.

The fix is to add all libraries to the `_allLibraries` field, not taking the
'include' list into consideration. Then when we decide what libraries are
considered to be part of a package, we filter down to 'included' libraries.