Improve logic of calculating a "fully qualified name" (#3836)
For improvements that I'm working on in another CL, it made sense to tidy up this code. The `_buildFullyQualifiedName` function was unnecessarily recursive and complex. And then the `_fullyQualifiedNameWithoutLibrary` function (a) has a complex name and (b) also worked _further_ to do undo some work done in `_buildFullyQualifiedName`. This CL tidies these two "fields".

* Combine `fullyQualifiedName` and `_buildFullyQualifiedName` into one getter
  (removing one field).
* Combine `fullyQualifiedNameWithoutLibrary` and
  `_fullyQualifiedNameWithoutLibrary` into one late final field, remove
  recursion, and simplify name to `qualifiedName`. Also mark as
  `@visibleForOverriding`. It should be private but because of the mixin
  hierarchy, it is declared in one file, and defined in another.
* Improve documentation.