Refactor "could apply" logic for extensions
This logic was implemented in a few places between Extension and ElementType (and subtypes). We don't need any logic to live in ElementTypes though, since any calculations that need to be made regarding DartTypes or InterfaceTypes can be made in Extension (`alwaysApplies` and `couldApplyTo`). This fixes https://github.com/dart-lang/dartdoc/issues/3846 and dramatically simplifies the implementation. Also:

* Rename `Extension.extendedType` to `Extension.extendedElement`.
* Change `PackageGraph._extensions` to a Set, to remove duplication in multiple canonical libraries.
* Introduce a classes test.
* Delete some redundant end2end testing code.