Fix many library sidebars, and search
The library sidebars on container (e.g. Class), top-level variable, top-level
function, and typedefs were totally busted, as a result of
d629e1eacccf11338413d2581818cc9c84f340d9.

Fixes https://github.com/dart-lang/dartdoc/issues/3859 and fixes
https://github.com/dart-lang/dartdoc/issues/3861.

The fix is to allow a Library's `canonicalModelElement` to be it's canonical
library, instead of `null`. This involves adding a few checks for "am I
calculating something for a Library?" and moving some "is this library
considered public" logic out of `Element.hasPrivateName` and into
`ModelElement.isPublic`. That corrects the issue of an exported Class's (for
example) enclosing element's URI, for the search results.

Once all of that is fixed, we also correct the `aboveSidebarPath`
implementation for a few elements, to point to the _canonical_ library's
sidebar path.

This fix highlights an issue with wildcard-variables, so one test is newly
marked as failing.