Correct the wildcard tests for import prefixes.
An unrelated fix, https://github.com/dart-lang/dartdoc/pull/3865, revealed that
the wildcard-variables experiment was not actually enabled in tests; the SDK
constraint used in test packages needs to be '^3.6.0'. With that fix, there are
some questions about how much dartdoc should be doing to keep wildcards
"non-binding" and how much the analyzer should be doing. So that is raised in
https://github.com/dart-lang/dartdoc/issues/3769#issuecomment-2327129506.