Support for enhanced parts
There isn't _much_ for dartdoc to be concerned with. Dartdoc basically must be
aware of an exports foud in a part, and there might be reasons to be concerned
with imports coming from a part. In particular, dartdoc used to skip resolving
all part files when searching for libraries, just as an optimization. Can't do
that any more.

This might not be a full implementation of enhanced parts support, but it sets
up the basics.